fix: aicard moved to community folder
