+++ src/community/ai-card.css.ts
@@ -1,7 +1,7 @@
 import {globalStyle, style, keyframes} from '@vanilla-extract/css';
-import {sprinkles} from './sprinkles.css';
-import {vars} from './skins/skin-contract.css';
-import * as mq from './media-queries.css';
+import {sprinkles} from '../sprinkles.css';
+import {vars} from '../skins/skin-contract.css';
+import * as mq from '../media-queries.css';
 
 const fill = (color: string) => `linear-gradient(${color}, ${color}) padding-box`;
 const border = `linear-gradient(200deg, var(--vivoPurple600, rgba(174, 66, 228, 0.35)) 17.51%, var(--vivoPurple500, rgba(189, 74, 255, 0.35)) 38.3%, var(--vivoPink500, rgba(235, 60, 125, 0.35)) 82.5%) border-box`;
