+++ src/community/index.tsx
@@ -1,3 +1,4 @@
+export {default as AiCard} from './ai-card';
 export {default as ExampleComponent} from './example-component';
 export {default as AdvancedDataCard} from './advanced-data-card';
 export {
