+++ src/index.tsx
@@ -101,7 +101,6 @@ export {default as Divider} from './divider';
 export {Menu, MenuItem, MenuSection} from './menu';
 export {default as EmptyState} from './empty-state';
 export {default as EmptyStateCard} from './empty-state-card';
-export {default as AiCard} from './ai-card';
 export {default as Callout} from './callout';
 export {default as Avatar} from './avatar';
 export {default as Slider} from './slider';
@@ -275,6 +274,7 @@ export type {TextToken, Dictionary} from './text-tokens';
  * the community.js export has issues because it exports an ES module and next12 interterprets it as a CommonJS module
  * importing from /dist/ is not an option because those modules don't get the context from the theme provider
  */
+export {default as CommunityAiCard} from './community/ai-card';
 export {default as CommunityExampleComponent} from './community/example-component';
 export {default as CommunityAdvancedDataCard} from './community/advanced-data-card';
 export {
